Add files via upload
This is my Power Panel widget. It displays information about power consumed during the flight.
Select the sensor and set the total usable capacity (divided by 100) in the set up screen.
The capacity only ranges from 0 to 100 when scrolling on the Horus Simulator, so I have set it up to use 35 = 3500 mAh, etc.
Set the total used to your preference. Leave in a buffer and land at zero or don't and land at 20% or whatever, up to you.